+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog
 
+### 0.7.3
+- **Update**: Xcode 9.3 migration.
+- **Remove**: Default initializer for Network service that conforms to `ConfigurableNetworkService` protocol.
+- **[Breaking Change]**: `DateFormattingService` class replaced with protocol.
+- **Add**: `SwiftDate` dependency for `DateFormattingService`.
+- **Add**: `ViewBackground` enum that describes possible view backgrounds.
+- **Add**: `ViewText` enum that describes text with appearance options.
+- **Removed**: `String+SizeCalculation` extension.
+
 ### 0.7.2
 
 - **Fixed**: Change root controller for window
+
 
 ### 0.7.1
 - **Add**: Extension for comparing optional arrays (`[T]?`) with `Equatable` elements.